Inline team general settings actions
